code/chapter13/helloagents-trip-planner/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.0",
    "fastapi>=0.115.0",
    "fastmcp>=2.0.0",
    "hello-agents[protocols]>=0.2.4,<=0.2.9",
    "httpx>=0.27.0",
    "huggingface-hub>=0.25.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "uv>=0.8.0",
    "uvicorn[standard]>=0.32.0",
]
